refactor: update structure, domain (User, Game)
