+++ src/main/java/baseball/domain/User.java
@@ -1,0 +1,13 @@
+package baseball.domain;
+
+public class User {
+    int[] userNumbers;
+
+    public int[] getUserNumbers() {
+        return userNumbers;
+    }
+
+    public void setUserNumbers(int[] userNumbers) {
+        this.userNumbers = userNumbers;
+    }
+}
